CORE-7863 Fix code reloading problem
This commit fixes a problem related to code reloading
Steps to reproduce:
1) rails s; go to any OnApp interface page
2) Add to any controller: puts '1'
3) reload page
Result:
'no method #authorized'

After a research I realize the problem was in onapp-core initializer.
We had there code like this:
Core::ApplicationController.include(ApplicationControllerMethods)
And it worked just fine with no changes in code. But as soon as some
change added to ActionPack code it was trying to reload all models. But,
unfortunately, it does not reload code in initializers. Thats why
User model located in onapp-core gem became not extended with :base,
:authorization and other concerns. As well as
Core::ApplicationController didn't include ApplicationControllerMethods.

The decision was to put these extencions into class variables and
include modules inside model, not in initializers.